net: lib: cloud: merge cloud_backend.h into cloud.h
Merge cloud_backend.h header file into cloud.h.
